store risk score in database

+++ Risk_Analysis/api/models.py
@@ -4,7 +4,7 @@ from djongo import models
 class categories_average(models.Model):
 
     category_name = models.CharField(blank=True, null=True)
-    score = models.IntegerField(blank=True, null=True)
+    score = models.FloatField(blank=True, null=True)
 
     class Meta:
         abstract = True
@@ -12,7 +12,7 @@ class categories_average(models.Model):
 class Time(models.Model):
 
     year = models.IntegerField(blank=True, null=True)
-    score = models.IntegerField(blank=True, null=True)
+    score = models.FloatField(blank=True, null=True)
 
     class Meta:
         abstract = True
@@ -20,7 +20,7 @@ class Time(models.Model):
 class User(models.Model):
 
     name = models.CharField(max_length=60,blank=True)
-    risk_score = models.IntegerField(default=None,blank=True)
+    risk_score = models.FloatField(default=None,blank=True)
     category_avg = models.ArrayField(model_container=categories_average,null=True,blank=True)
     risk_run = models.ArrayField(model_container=Time,null=True,blank=True)
     objects = models.DjongoManager()
